+++ sdk/core/azure-core/test/nlohmann-json-test/src/unit-comparison.cpp
@@ -29,7 +29,7 @@ SOFTWARE.
 
 #include "doctest_compatibility.h"
 
-#include <azure/core/internal/json.hpp>
+#include <azure/core/internal/json/json.hpp>
 using Azure::Core::Json::_internal::json;
 
 namespace {