+++ sdk/storage/azure-storage-common/inc/azure/storage/common/storage_switch_to_secondary_policy.hpp
@@ -6,7 +6,7 @@
 #include <memory>
 #include <string>
 
-#include <azure/core/http/policy.hpp>
+#include <azure/core/http/policies/policy.hpp>
 
 namespace Azure { namespace Storage { namespace _detail {