+++ sdk/storage/azure-storage-files-datalake/src/datalake_file_client.cpp
@@ -3,7 +3,7 @@
 
 #include "azure/storage/files/datalake/datalake_file_client.hpp"
 
-#include <azure/core/http/policy.hpp>
+#include <azure/core/http/policies/policy.hpp>
 #include <azure/storage/common/constants.hpp>
 #include <azure/storage/common/crypt.hpp>
 #include <azure/storage/common/shared_key_policy.hpp>